pyproject.toml
[project]
name = "sbs-ext"
version = "0.1.0"
description = "Extensions for Smart Book Standard (SBS)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "markdown-it-py[plugins]",
    "pytest",
    "pyyaml",
    "uvicorn",
]
